pyproject.toml
[project]
name = "zentra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.2",
    "django>=6.0.3",
    "django-celery-beat>=2.9.0",
    "django-cors-headers>=4.9.0",
    "django-debug-toolbar>=6.2.0",
    "django-filter>=25.2",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-nested-routers>=0.95.0",
    "drf-spectacular>=0.29.0",
    "factory-boy>=3.3.3",
    "faker>=40.8.0",
    "gunicorn>=25.1.0",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "python-dateutil>=2.9.0.post0",
    "python-decouple>=3.8",
    "redis>=7.2.1",
    "setuptools>=82.0.0",
    "weasyprint>=68.1",
    "wheel>=0.46.3",
    "whitenoise>=6.12.0",
]
